+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support for retrieving various Model Derivative data as readable stream
+
 ## [8.0.7] - 2019-11-20
 
 ### Added